requirements.txt
requests
sentence-transformers
numpy
flask
flask-cors
openai
groq
python-dotenv